+++ src/backend/PeopleHub.Application/Models/FeedPost.cs
@@ -1,3 +1,3 @@
 namespace PeopleHub.Application.Models;
 
-public sealed record FeedPost(long Id, string Text, int AuthorUserId);
+public sealed record FeedPost(long Id, string Text, long AuthorUserId);
