+++ src/backend/PeopleHub.Domain/Entities/FriendRequest.cs
@@ -3,32 +3,32 @@ using PeopleHub.Domain.Exceptions;
 
 namespace PeopleHub.Domain.Entities;
 
-public sealed class FriendRequest(int id, int senderUserId, int receiverUserId, FriendRequestStatus status)
+public sealed class FriendRequest(long id, long senderUserId, long receiverUserId, FriendRequestStatus status)
 {
-    public int Id { get; } = id;
-    public int SenderUserId { get; } = senderUserId;
-    public int ReceiverUserId { get; } = receiverUserId;
+    public long Id { get; } = id;
+    public long SenderUserId { get; } = senderUserId;
+    public long ReceiverUserId { get; } = receiverUserId;
     public FriendRequestStatus Status { get; private set; } = status;
 
-    public static FriendRequest Send(int senderUserId, int receiverUserId)
+    public static FriendRequest Send(long senderUserId, long receiverUserId)
     {
         EnsureDifferentUsers(senderUserId, receiverUserId);
         return new FriendRequest(0, senderUserId, receiverUserId, FriendRequestStatus.Sent);
     }
 
-    public static FriendRequest EstablishFriendship(int userId, int friendUserId)
+    public static FriendRequest EstablishFriendship(long userId, long friendUserId)
     {
         EnsureDifferentUsers(userId, friendUserId);
         return new FriendRequest(0, userId, friendUserId, FriendRequestStatus.Approved);
     }
 
-    public void Approve(int byUserId)
+    public void Approve(long byUserId)
     {
         EnsureReceiver(byUserId);
         Status = FriendRequestStatus.Approved;
     }
 
-    public void Reject(int byUserId)
+    public void Reject(long byUserId)
     {
         EnsureReceiver(byUserId);
         Status = FriendRequestStatus.Rejected;
@@ -36,7 +36,7 @@ public sealed class FriendRequest(int id, int senderUserId, int receiverUserId, 
 
     public void MarkApproved() => Status = FriendRequestStatus.Approved;
 
-    private void EnsureReceiver(int userId)
+    private void EnsureReceiver(long userId)
     {
         if (userId != ReceiverUserId)
         {
@@ -44,7 +44,7 @@ public sealed class FriendRequest(int id, int senderUserId, int receiverUserId, 
         }
     }
 
-    private static void EnsureDifferentUsers(int senderUserId, int receiverUserId)
+    private static void EnsureDifferentUsers(long senderUserId, long receiverUserId)
     {
         if (senderUserId == receiverUserId)
         {
