+++ src/backend/PeopleHub.Domain/Entities/User.cs
@@ -5,7 +5,7 @@ namespace PeopleHub.Domain.Entities;
 
 public sealed class User(int id, PersonalInfo personalInfo)
 {
-    public int Id { get; } = id;
+    public long Id { get; } = id;
     public PersonalInfo PersonalInfo { get; private set; } = personalInfo;
 
     public static User Create(PersonalInfo personalInfo) =>
