+++ src/backend/PeopleHub.Infrastructure/Caching/IFeedCacheService.cs
@@ -4,6 +4,6 @@ namespace PeopleHub.Infrastructure.Caching;
 
 public interface IFeedCacheService
 {
-    Task PushFeedAsync(int userId, IReadOnlyCollection<FeedPost> feedPosts);
-    Task<IReadOnlyCollection<FeedPost>> GetFeedAsync(int userId);
+    Task PushFeedAsync(long userId, IReadOnlyCollection<FeedPost> feedPosts);
+    Task<IReadOnlyCollection<FeedPost>> GetFeedAsync(long userId);
 }
